[zircon] Benchmark full VMO lifecycle
This change adds a new benchmark test to measure the cost of a full VMO
lifecycle.  This is useful because this is essentially what users of
`fuchsia.mem.Data` or `fuchsia.mem.Buffer` must do on top of their
default zx.channel write/read operations.  Thus, the sum of those two
metrics gives a good idea of the cost of a sending those two FIDL types.

Test: fx test --device -ov fuchsia_microbenchmarks
Bug: 100478
Change-Id: I4cf946f491361b8e422176e1eb3f69a7c3c32616
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/684173
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>
Reviewed-by: Mark Seaborn <mseaborn@google.com>