[garnet][build] Move //garnet/packages build files
This moves the BUILD.gn files from //garnet/packages to
//bundles/packages to enable removing the //garnet source tree.

One rename of a group is made to avoid using the word "garnet". This is
change //garnet/packages:garnet to //bundles/packages:base_packages.

This is a "soft" transition since there are projects outside fuchsia.git
which depend on some of these groups. This CL will create them, and then
the external dependencies can be updated to these, then finally remove
the copies in //garnet.

Bug: 36063, 101982

Change-Id: Ibccdd481a64ff8f4aa2b8da3b807a02f862b3999
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/687982
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>