[ktrace] Perform CopyToUser untracked
User copying is not generally allowed to be performed with locks held.
For ktrace lock_ is acting specifically to serialize reads with other
operations, and so makes sense to hold whilst copying. This change
removes tracking for the duration of the copy to prevent the
AssertNoLocksHeld validation from tripping.

Bug: 101783
Change-Id: Iaaaa1832da5e93d27f3e61c06e12629cfc892e9c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/687826
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: John Grossman <johngro@google.com>