[starnix] Mount the galaxy and running package inside the galaxy filesystem
This CL ensures that the galaxy package is always mounted at /galaxy in
a running galaxy. Each component that then needs access to its package
because its executable is inside it will be mounted in
/galaxy/pkg/{HASH} where {HASH} is the sha-256 of its /meta/content
file.

For this to be possible a new layered filesystem is created. This
filesystem will delegate to a base one, but have a number of top level
directory that points to other filesystem. This allows to layer the root
of the galaxy image the following way:

LayeredFs:
  base: system image
  galaxy: LayeredFs:
    base: galaxy package
    pkg: tmpfs

Change-Id: Iad52f1e7943fca6e098052d8e3795aefa4d7b43b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/685762
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Adam Lesinski <adamlesinski@google.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>