[fidl] Rust measure tape GIDL tests
This CL adds GIDL support for generating measure tape tests for Rust.
While this only covers Rust, much of the calculation logic is shared
with the HLCPP measure tape.

This CL reuses the existing measure tape target-types argument, meaning
that all types are listed out. This is done for the sake of simplicity
and the fact that both the allow lists and denylists are currently
both long.

Bug: 101585
Bug: 101893
Change-Id: I04d27ea81939cb9ec1aba54991faa0b972bc6eba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/685688
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>