[starnix] Refactor DirentSink implementations
This CL moves the responsibility of updating the file position inside
the implementation of DirentSink.

This ensures that the position is kept in sync with the value returned
to user space, and ensures that the file position lock is held uniformly
during the call to readdir across all filesystems implementation.

Change-Id: I3084e13e5ac6789aa73c10031161980aefb3548c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/688407
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Adam Barth <abarth@google.com>