[component_manager] s/Purge/Destroy
Purge and Destroy are now unified, change the naming to reflect as such.

Exceptions for places that still rely on the Purged event, including hub
and external clients. These are a bit more involved than a simple rename
so we'll take care of them separately.

Bug: 100652
Change-Id: I9596434b71bb00f8aee799eb6ca3d2af3d4fa005
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/687844
Reviewed-by: Justin Mattson <jmatt@google.com>
Commit-Queue: Gary Bressler <geb@google.com>