[emu] Fix ffx config data for emulator
During recent troubleshooting, we found that the ffx config value to
control the startup timeout for `ffx emu` was incorrectly populated in
the config.json file. The name of the variable is meant to be
emu.start.timeout, with a default value of 60. Instead it was being set
as emu.startup.timeout, with a default of 60.0. This CL sets the correct
default values.

Change-Id: Id8379ffd8c9312430cb49e1b3724f07a3b8b7a81
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/688684
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>