[libc] Clear start_arg field from pthread struct
Currently, a pointer passed via the start_arg field is accessible during
the entire lifetime of a thread. This can hide an actual leak if the
start_arg was allocated data that a newly created thread never freed.
This CL clears the field right before entering the thread. This pointer
should be available in either the new thread's stack or registers (or
globals).

Bug: 66819
Change-Id: I38a2c24c48434b74770b173d1d697d5bf52d8292
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/671534
Commit-Queue: Leonard Chan <leonardchan@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>