[ui][text] Remove text_manager from//src/ui:runtime
Having text_manager be there is an excellent aspirational goal to have.
However, due to unintended interactions with the resolver, this makes
text_manager not be in "base" image in some products. We can't do
that, so to try and fix things, we do the easy stuff first.

The effect of this change should be to move text_manager back to 
"base".

Change-Id: I835286d2ab28de492b5c52d0d04db2dcd66179b1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/688622
Reviewed-by: David Worsham <dworsham@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>