[assembly][security] Disable scrutiny verification on boards using custom signing scripts
Disable scrutiny's checks that require parsing the zbi, if the board
uses a custom_signing_script.

A custom signing script is meant to be used when the zbi image must be
post-processed in some manner before it can be used by the target
device.  In those cases, the resultant file is likely not readable by
scrutiny, and so these checks do not work as written.

Bug: 100989

Change-Id: I71846d6eed0046ef3d069653d20e6d6a882ea48f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/682824
Reviewed-by: Allison Pearce <ampearce@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Mark Dittmer <markdittmer@google.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>