[vaapi] Exposing Y tiling format to client
Allow the VAAPI decoders to expose the underlying Y-tiling format that
Intel platforms process surfaces in. This allows for a zero copy approach
where we will be able to share the output of the DPB directly with our client
and not have to untile the data.

Bug: 99091, 81304
Change-Id: I19be7c277ce2532444db596bd920e2e80b13ec82
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681765
Reviewed-by: John Bauman <jbauman@google.com>
Commit-Queue: Stefan Bossbaly <stefanbossbaly@google.com>
Reviewed-by: Dustin Green <dustingreen@google.com>