[fidl] Convert ethertap to LLCPP
This is an attempt at constructing a minimal CL that removes C binding
usage from ethertap and moves it over to LLCPP.

In several parts of the CL, it may be possible to move to more ergonomic
usage of LLCPP but this was avoided to more quickly remove usage of the
C bindings and because some quick attempts at simplifying things
resulting it hanging tests that needed further debugging.

The original motivation for creating this CL was to remove usage of
fidl::Buffer.

Bug: 101890
Change-Id: I931c4c02820bd2edc9e2c404c9efc0099440b307
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686972
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>