[display][i915] Simplify pipe attachment logic.
The display-pipe mapping never changes once the pipe is attached,
until the display is disconnected; so we can simplify AttachPipe()
by not allowing assigning a display to another pipe. Also the pipe
mode is not set when calling AttachPipe(), so the modeset code in
that function will be never run and we can just remove it; pipe will
be initialized on the first ApplyConfig() where it is involved.

Bug: 102095
Change-Id: I8c1718d03856121b15c935cdf63b04a7b79e43b1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/688908
Reviewed-by: Randall Bosetti <rlb@google.com>
Commit-Queue: Yilong Li <liyl@google.com>