[rust][build] Use rustc_prefix everywhere
We previously had two build args, rustc_prefix and rust_sysroot, which
you never wanted to change separately as far as I can tell. The
inconsistent naming was confusing, especially because we use
rust_sysroot to mean something entirely different (the linker sysroot
used when linking Rust programs) elsewhere in the build.

This changes rustc_prefix to point to where rust_sysroot used to point,
which is the root directory of the rust package instead of the bin
directory that contained rustc. This matches the usual meaning of
"prefix" in my experience. The change means that existing usages of
rustc_prefix had to be replaced with `$rustc_prefix/bin`. In some cases
it simplified things because we were doing things like
`$rustc_prefix/../lib`, which is now just `$rustc_prefix/lib`.

Change-Id: I55a9ebf07b5c8ea10d5cecdf54bf2c8c8732e12a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/676422
Reviewed-by: Dan Johnson <computerdruid@google.com>
Commit-Queue: Tyler Mandry <tmandry@google.com>