[f2fs] Fix a deadlock problem in FileCache
This CL addresses a deadlock problem in FileCache::GetPage(). It could
happen when a thread was waiting for the lock of the n-th Page to be
released after acquiring the tree lock while another thread tried to
acquire the tree lock to get the n+1th Page with the n-th Page locked.
It resolves the deadlock by releasing the tree lock while a thread is
waiting for a locked Page in FileCache::GetPage().

Fixed: 100995

Change-Id: Ib61ccf18cce5adda7594b7536534363d44728d26
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/688969
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aaron Drew <ripper@google.com>
Reviewed-by: Aaron Drew <ripper@google.com>