[ffx] Add pretty_printed aggregated memory profile.
When called with no other arguments, `ffx profile memory` will now
print a human readable digest of the system's memory status. The output
closely matches that of `fx mem`, in the sense that it presents the
same information in the same order, but not necessarily with the same
formatting.

This is intended as a direct port of `fx mem`, and should be more or
less bug-compatible with it.

Bug: 231425139
Test: fx test ffx_profile_memory_lib_test
Change-Id: I8e08decafe7d0a4662cc50bdbf525edf446f4e15
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/683506
Reviewed-by: Jean-François Geyelin <jif@google.com>
Commit-Queue: Guillaume Chelfi <chelfi@google.com>
Reviewed-by: Pierre Labatut <plabatut@google.com>