[input][test] store event received from client to vector
Currently, we use 1 time usage callback to receive and validate
events from client (Flutter/Chromium). It does not work for the
WebEngine (Chromium) tests, because:

1. WE may not ready to process events after WE start.
2. Test retry sending events to WE and wait until the event
   processed by JS and response to test. (In the following
   CL: fxrrev.dev/665843).
3. For the retry, test does not know how many event maybe
   processed by JS. For example, for a click, the first mouse
   down may or may not received by WE. But current 1 time
   usage callback can only handle a fixed number of events.

In this cl, we store ResponseListenerServer received events to
vector, then tests can wait on the state of vector, eg. size of
vector, last event of the vector.

Bug: 94072
Change-Id: I5a32973e62b469f301039564f35f4c6a71a7552e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/688428
Reviewed-by: Caroline Liu <carolineliu@google.com>
Commit-Queue: Jianpeng Chao <chaopeng@google.com>