[fidl] Move dyn suite control channel to natural types
Natural types are recommended for most users, so we might as well use it
for our test harness. A potential risk of this is that there might be a
failure that effects the natural bindings but not the LLCPP bindings,
but they share most functionality other than the type definitions so it
shouldn't be too much of a risk. Also, most base functionality should be
covered elsewhere.

Change-Id: I050a1bf6ba071740537592b64b8d97b07655e1ad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/689582
Fuchsia-Auto-Submit: Benjamin Prosnitz <bprosnitz@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Yifei Teng <yifeit@google.com>