Revert "[guest_manager] Remove deprecated dependency"
This reverts commit a40bdbf44ca90782081f5c978f25e8fc12736102.

Reason for revert: This breaks Linux Terminal on workstation. I believe
this is because the previous solution provided a flat directory and the
updated code placed services in a nested svc directory, which breaks
service resolution.

The revert is not pure as the ServiceProviderBridge has been moved. The
only change was to fixup that path and add guest_manager to the
visibility list.

Note: this component will be removed shortly as part of CFv2 migration,
so I don't plan on rolling forward a different fix at this time.

Bug: 102111
Change-Id: Iaefb0025eb47e5b6c87488f8725f4bf199cb968c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/689722
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Prashanth Swaminathan <prashanthsw@google.com>
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>