[fidl-async-2] Remove server functionality
There does not appear to be any users.

`fidl::BindServer` should be able to cover the majority of use cases.
It's missing one which is putting a hard limit on the number of
concurrent transactions. We could consider adding that feature if
needed.

Change-Id: I31b5d3a7f6d5a288adf7b046ce745e42b3d8f3c9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/687123
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Yilong Li <liyl@google.com>
Reviewed-by: Dustin Green <dustingreen@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>