[Symbolizer] Clarify format specifiers in spec.
This change makes the meaning of "%i" explicit and removes the implicit
dependency of the format specifiers on C printf/scanf.

The "%x" used as the size argument in mmap is also changed to "%i", to
align with actual usage ("0x" is typically present, but disallowed by
"%x").

Change-Id: Ifb1c65369e5109ce707a604d7c281b06badd9a16
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/689107
Commit-Queue: Daniel Thornburgh <dthorn@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>