[flatland] Don't post task on link resolution.
In fxr/551822 and fxr/553362 we modified ObjectLinker::Link to
optionally post a link resolution task if a dispatcher is specified.
This was used to ensure that responses to ParentViewportWatcher and
ChildViewWatcher are run on the corresponding Flatland session's
thread (as required by the HLCPP FIDL bindings we *were* using).

However, this resulted in flaky behavior, as chronicled in fxb/100250.
Now that we have switched to the LLCPP FIDL bindings, we can fix the
flake by passing nullptr instead of the optional dispatcher.

Fixed: 100250
Change-Id: I469aa4020ebade5104e7235a6e1e49c3b67f5971
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/688903
Reviewed-by: Emircan Uysaler <emircan@google.com>
Commit-Queue: Josh Gargus <jjosh@google.com>