[netstack3] Return socket state on failed insert
When an insert into BoundSocketMap fails, return the state that was to
be inserted instead of dropping it. This will make it easier to support
multicast registrations, which should be restored with the unbound
socket state after a failed bind.

Change-Id: Ib7d8384a041433cfb83c0c5c32cf60a3f0ff4932
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/690142
Fuchsia-Auto-Submit: Alex Konradi <akonradi@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>