[fint] Fix bug in affected BUILD.gn file analysis
When checking whether a test was affected based on whether its BUILD.gn
file was changed, we were previously comparing the _absolute_ path of
the BUILD.gn file ith the _relative_ path of the test's GN label, so we
were never considering any tests to be affected by their GN file
changes.

Now we always use relative paths to compare test labels to changed
files.

Also simplify the `removeDuplicates()` logic as long as I'm here. It was
optimized to avoid unnecessary allocations, but that made it
significantly harder to read and made me suspect that it was concealing
the bug that was causing this issue. It will never be used to sort more
than a few thousand elements at a time, so the optimization was not
worth the loss of readability.

Bug: 101789
Change-Id: Ie07ed4f113f994e86755981ee1f98733190be183
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/690149
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Shai Barack <shayba@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>