[kernel] More precisely conditialize '64-bit' printf test cases
We were incorrectly relying on the presence of the _LP64 preprocessor
variable to establish whether certain integral types are 64-bit (e.g.,
size_t). In actuality, this variable is only defined when `long int`
and pointers are 64-bit _and_ `int` is 32-bit - and so is a poor test of
whether, say, size_t is 64-bit. As a concrete counterexample, the x86
EFI target has `long int` as 32-bit, but `size_t` as 64-bit - so in
particular, this change allows the printf tests to pass in that
environment.

Change-Id: I9ac24d57cb1c3b3904754cd7266bfe493cfc0f78
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/690162
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>