[vaapi] Share tiled DPB directly with clients
Instead of copying DPB data to VMOs have the DPB be backed directly by
VMOs in the tiled case. This way we can share directly the VMOs with the
client and not have to copy. To accomplish this and maintain
compatibility with the linear format, a SurfaceBufferManager interface
was introduced. Each version of this interface must handle the
relationship between packets, buffers, DPB and VMOs.

Bug: 99091

Change-Id: I4f5f0781a2ba2731a310e75445f994d6c455be28
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/682363
Commit-Queue: Stefan Bossbaly <stefanbossbaly@google.com>
Reviewed-by: Dustin Green <dustingreen@google.com>