[netemul] Ensure unique ethertap endpoint names
...by appending a unique integer ID to them.

Because device cleanup is asynchronous for ethertap devices, it's
possible for device cleanup to race with the creation of a new device
with the same name, leading to a name collision. To avoid the collision,
"salt" the endpoint names.

Fixed: 102220
Change-Id: Ia034f149ea9f55e6df53412b90f29f0c1965bafb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/690365
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>