[docs][style] Revert change to line width.
The line width of 80 was changed to 100 in
https://fuchsia-review.googlesource.com/c/fuchsia/+/498703
which misunderstood the difference as an oversight.

I'd make a revert for that here
https://fuchsia-review.googlesource.com/c/fuchsia/+/682525
but the file has changed to where a simple revert no longer applies.

So this CL is a direct change, along with more explanation (to help
keep this from changing again). But in some sense, this is still a
kind of 'revert' of 498703.

Change-Id: Id22cff853458f603ec911f8a1d570d90d9d15045
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/690522
Commit-Queue: Dave Schuyler <dschuyler@google.com>
Reviewed-by: Vivian Tran <viviantran@google.com>