[fidl][rfc] RFC-0138 change error handling.
Change how RFC-0138 defines error handling for unknown interactions,
both in the wire format and in the syntax.

In the wire format: define `fidl.TransportErr`, for the `transport_err`
variant of the result union, in order to avoid dependencies on library
`zx` and to clearly define how invalid values of `transport_err` are
handled.

In the bindings: specify that error handling should fold `transport_err`
into existing error handling for transport-level errors such as errors
from channel read/write and decoding, and that bindings should actually
present the same API for strict or flexible methods, rather than
presenting an additional transport error variant.

This is based on experience while implementing the RFC: LLCPP, C++
Natural Types, and Rust bindings all use similar result types, which
present transport errors and application errors in a way that can be
differentiated and handled, and given the rest of the design here, it
made more sense to roll into the existing transport errors rather than
attempting to add another layer of complexity to the result types.

The RFC when drafted was likely focused on HLCPP, where all transport
errors are sent to a separate handler, which would require using a
different API for flexible methods than strict methods in order to let
bindings tell which method call got an unknown interaction response.

The originally proposed approach will likely still be used if unknown
interaction support is implemented in HLCPP. However, by changing the
RFC to match the usage in LLCPP, C++ Natural Types, and Rust, we make it
so those bindings are the norm and HLCPP is the exception, which is
preferred since HLCPP is likely to be deprecated.

Bug: 88366
Change-Id: Ied132b8d5d296908303b5a7bd8779460abf6b88d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686966
Commit-Queue: Zachary Stewart <zstewart@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Adam Barth <abarth@google.com>