[mdns] remove additional empty string from generated TXT records
This CL fixes a bug in the DNS message writing code that added an
empty text string to the list of text strings in a TXT resource
record. An empty text string is still added when there are no other
strings in the resource.

Fixed: 102543

Test: new unit tests
Change-Id: I2ba89ae6db1b244f5e87da19a9f31327b05a98e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/690412
Reviewed-by: Sundarajan Srinivasan <sundarsrini@google.com>
Commit-Queue: Dale Sather <dalesat@google.com>