[wlan][policy] Explicitly select authentication method.
This change constructs an `Authentication` FIDL message in the Policy
client state machine when connecting to networks. To accomplish this,
the configured security protocols of scanned networks is included in
`ConnectionCandidate` and used to select an authentication method.

Some information is only present in a `ConnectionCandidate` if a scan
has been performed. Since this change introduces more of this kind of
data, these fields have been refactored into `ScannedCandidate` and a
single optional field of this type.

The `Authentication` FIDL message is not yet used and this change
temporarily discards everything but the bare credential.
`Authentication` will be used directly following changes to the SME FIDL
APIs at which point this temporary code will be removed.

Bug: 72051
Bug: 72906
Testing: Unit tests and manual testing.
Change-Id: I2e4d8f8e85346c02d27daf985bcf231326063c35
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/672744
Commit-Queue: Sean Olson <seanolson@google.com>
Reviewed-by: Marc Khouri <mnck@google.com>