[docs] Fix header names for syscalls.
The NAME header is for a section containing a summary
of the syscall. It does not contain a name.

The SYNOPSIS header is for a section containing the
C function signature of the syscall. It does not
contain a synopsis.

Changed NAME -> SUMMARY.
Changed SYNOPSIS -> DECLARATION.

Change-Id: I1ad8878836d86e0570d1befdcfd7f034ba2d4361
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/689004
Reviewed-by: Vivian Tran <viviantran@google.com>
Commit-Queue: Alexander Biggs <akbiggs@google.com>
Reviewed-by: Nick Van der Auwermeulen <nickvander@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Theodore Siu <theosiu@google.com>