[fxfs] Roll the journal mutations key
The Chacha20 variant we are using only allows a 32-bit block
offset (where blocks are 16 bytes) which isn't so big that we need to
not worry about wrapping.

To address this, this CL will roll the mutations key after it has been
used for 128 MiB.

In addition to this, this CL tidies up some other encryption related
issues:

 1. The encryption keys are now a separate key/value pair which avoids
    having to update it every time the allocated size or reference count
    changes.

 2. Moved the key ID out of WrappedKey and UnwrappedKey since that is
    specific to keys used for extents.  WrappedKeys and UnwrappedKeys
    now store tuples.  This means that we can use WrappedKey and
    UnwrappedKey consistently.

 3. EncryptedMutations can be bigger than the default maximum
    serialization size allowed.  It used to fail if we exceeded that
    limit.  This CL fixes that.

As this is a breaking change, I've removed support for all older
versions.

Fixed: 93354
Change-Id: I4c43dff1fabbed084f66c608799ead584b4a75bf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/685705
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Aaron Drew <ripper@google.com>
Commit-Queue: Chris Suter <csuter@google.com>