[net-cli] Make net route --metric optional
The fuchsia.net.stack.ForwardingTable FIDL definition documents a
`metric` of 0 to mean that the metric is unspecified:
https://cs.opensource.google/fuchsia/fuchsia/+/main:sdk/fidl/fuchsia.net.stack/stack.fidl;l=28;drc=2707f1db129a8c6fa20b5a5aa282fde9dcd400dd

We should expose this fact in `net-cli` for the convenience of
network-conformance testing scripts. `net-cli` used to expose access
to an equivalent `fwd` command which did allow for omitting the metric,
but the `fwd` command was removed in
https://fuchsia-review.googlesource.com/q/I71326cd4c894f88e2d4aae524433e69c853c532a.
I overlooked the fact that `metric` was made non-optional when migrating
testing scripts to use `route` instead of `fwd`.

Test: net-cli-tests
Change-Id: I5daf8a9a1bbe63f38f4545deff5cf193a915bca6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/690822
Reviewed-by: Jake Fried <friedj@google.com>
Reviewed-by: Nick Brown <nickbrow@google.com>
Commit-Queue: Evan Wildenhain <wildenhain@google.com>
Fuchsia-Auto-Submit: Evan Wildenhain <wildenhain@google.com>