[netstack3] Extract device related ops to non-sync ctx
This change extracts the all contexts that operate on a device to the
non-synchronized context. This includes the `EventContext`s (which all
emit events related to device changes) and `DeviceLayerEventDispatcher`
(which sends frames).

The change had to do these together instead of extracting `EventContext`
and `DeviceLayerEventDispatcher` in different changes because both need
access to the `bindings::devices::Devices` structure and if both were
not moved to the non-synchronized context together, we would need to
mutably access shared state, requiring locks.

Bug: 48578, 86168

Change-Id: I84b11f040721dcb36ff48ee84fee5d652684e375
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/690711
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jake Fried <friedj@google.com>
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>