[fidlc] Add compiler support for new-types
This CL adds support for the new-type syntax in the compiler. This
patch is responsible for adding the code that will handle new-type
declarations and a flag to gate this feature. The code for emitting
the correct JSON for the IR will be provided in a dependent CL.
Similarly, the language specific binding generation code will be
provided in separate CLs.

Bug: 7807
Change-Id: Ic5b5691f42ee79e2fe5ecfd9241d173cd640f301
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681402
Reviewed-by: Alex Zaslavsky <azaslavsky@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Mitchell Kember <mkember@google.com>