[build] Remove system_image allowed_in_extra_deps
None of the system packages require this flag, so remove it.
This flag lets system packages be included anywhere in the build,
so it makes the build graph harder to understand.

Now system packages have to be included in //build:input:system_image
or they will not show up in /system/.

Bug: 69738
Change-Id: I84dbbe6ae1362bad078359531fe403412f38ddf0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/688852
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Nick Eaton <nickeaton@google.com>