[fx][rbe] fx set --rust-rbe, use rust_rbe.enable
fx set now provides `--rust-rbe` as a renamed version of `--rbe`
(to be phased out).  `--rbe` continues to work (non-breaking).

Use the new variable `rust_rbe_enable`
as a replacement for the old `enable_rbe`, which will be removed
in the future.  We print a warning to update for now.

This will avoid confusion with a future `--cxx-rbe` flag.

Bug: 101917

Change-Id: I45299ff62fce9610e46c76ceb2c2a1bb75b6e2e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/688147
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Reviewed-by: Renato Mangini Dias <mangini@google.com>
Commit-Queue: David Fang <fangism@google.com>