[msd-arm-mali] Put GPU in nonprotected mode before perf counter test
The perf counter test expects the GPU to be in unprotected mode so it
can write out the performance counters. Some earlier tests may leave it
in protected mode, so we need to send a non-protected atom to the GPU to
force it to always switch to non-protected mode.

Fixed: 102538
Change-Id: Idadf074e9366fc05b58cb628903ee13ac3715bf1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691122
Reviewed-by: Craig Stout <cstout@google.com>
Fuchsia-Auto-Submit: John Bauman <jbauman@google.com>
Commit-Queue: John Bauman <jbauman@google.com>