[system-updater] Stop using pkgfs client library
system-updater now accesses its namespace directly and uses
fuchsia_pkg::PackageDirectory instead of pkgfs::system::Client
to get the hash of the system_image package. system-updater is the
last client of pkgfs::system::Client and the pkgfs client library
is the last client of pkgfs-ramdisk, so with this change both can be
deleted.

It is reasonable to get the system_image hash with custom
code instead of using a /pkgfs/system specific library because:
* This is the last SWD use of the system_image package outside of
  pkg-cache, and pkg-cache uses a type that wraps a
  package_directory::RootDir instead of a DirectoryProxy.
* There should not be any new uses of /pkgfs/system in the future,
  because only pkg-cache should be directly interacting with the
  system_image package. All non-pkg-cache uses of information
  contained in the system-image package should be obtained through
  pkg-cache-served FIDL, e.g.
  fuchsia.pkg/PackageCache.BasePackageIndex

Bug: 102560
Change-Id: I1579b042f64f9d04f7fb80608177746488499284
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/690546
Reviewed-by: Sen Jiang <senj@google.com>
Commit-Queue: Ben Keller <galbanum@google.com>
Fuchsia-Auto-Submit: Ben Keller <galbanum@google.com>