[identity] Delete TokenManagerFactory.
TokenManagerFactory was intended as a way for products to integrate
with Oauth and OIDC providers without using the full account system. We
don't have any products that currently need this capability nor do we
know of any in the future.

This 2018 implementation has not been maintained and would likely need
an overhaul before it could be used and the reality is that products
would be as likely to bring their own implementation as to use one
provided by the platform. Deleting it now saves the ongoing cost of
maintaining it and in particular the need to migrate it to CFv2.

Bug: 101940,101941
Change-Id: I48249465886ffe252850d91e2ed88aa5e94f0970
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/690673
API-Review: Shai Barack <shayba@google.com>
Reviewed-by: Angad Pal Singh <apsbhatia@google.com>
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>
API-Review: Dale Sather <dalesat@google.com>
Reviewed-by: Justin Mattson <jmatt@google.com>